New mainfile w/ CLI args, improved dict attack
Hello! I found this project while exploring GitHub
and decided to fork it.

Moved main program login to Main.py, I do plan on
renaming this eventually. Created Dictionary.py,
which handles dictionary attacks on zipfiles and
extracts them to <filename>.zip_cracked. Brute.py
is being worked on in the bf-mthd branch, which
will be merged upstream eventually. Created a "default"
dictionary (default.txt) that will eventually be
populated from a larger database of passwords - it
is just used for testing right now, but feel free to
add to it.